Enabling all warnings in all projects and only selectively disabling those that are needed
with aim to eventually update the code to need as little warnings disabled as needed.

Some code updates (and even fixes) to avoid warning done already

Source commit: 5c45a6f56c52e7b0c5034ed9980cd65259a7372b

+++ source/filezilla/AsyncProxySocketLayer.h
@@ -129,7 +129,7 @@ private:
   char *m_pStrBuffer;    // Recvbuffer needed by HTTP1.1 proxy
   int m_nProxyOpState;  // State of an operation
   int m_nProxyOpID;    // Currently active operation (0 if none)
-  int m_nProxyPeerPort;  // Port of the server you are connected to, retrieve via GetPeerName
+  unsigned short m_nProxyPeerPort;  // Port of the server you are connected to, retrieve via GetPeerName
   ULONG m_nProxyPeerIp;  // IP of the server you are connected to, retrieve via GetPeerName
   typedef struct
   {
